Modify the default Grafana log query to not have empty values
Loki queries are limitted to at least one `non-empty` conditional, but
since the current query in the default Application logs dashboard
uses variables that have `.*` regex as their `All` value, they are
capable to become fully empty when `All` is selected. And since loki
dislikes this possibilty it throws an error whenever all the variables
are chosen to be `All` which is also the first thing we see when we enter
the dashboard.

The fix is simple and is actually already implemented in
`grafana_dashboards/spring_boot.json` file: at least a single condition
that cannot be empty. My commit does a similar change by making all
the variables in the query such as `juju_application` expect to be
non-empty explicitly. Making this change to only a single variable
would also be sufficient, but I decided to make it explicit.